docs(react-native): add filters screen to example
