+++ packages/react-instantsearch/examples/react-native/index.ios.js
@@ -6,4 +6,5 @@
 
 import { AppRegistry } from 'react-native';
 import App from './App';
+
 AppRegistry.registerComponent('reactnative', () => App);
